+++ embassy-mspm0/src/gpio.rs
@@ -1209,7 +1209,7 @@ const DISCONNECT_PF: u8 = 0;
 /// The pin function for the GPIO peripheral.
 ///
 /// This is fixed to `1` for every part.
-const GPIO_PF: u8 = 1;
+pub(crate) const GPIO_PF: u8 = 1;
 
 /// A pin with wakeup logic, able to bring the device out of SHUTDOWN.
 ///